Refactor: some more synth stuff going in. Might give up on this because of sample rate issues.

+++ Framework/Audio/Synthesizer/Oscillators/IOscillator.cs
@@ -12,6 +12,6 @@
         /// <param name="frequency">The frequency.</param>
         /// <param name="volume">The volume. This is a value between -1 and 1.</param>
         /// <returns>The audio signal.</returns>
-        float GetSignal(float time, float frequency, float volume);
+        double GetSignal(double time, double frequency, double volume);
     }
 }
